test: put back four more checks from the review
- Agent: the conversation retention forwarder is registered. The daily job
  that deletes expired conversations asks for it, and that job is not in DI
  itself, so nothing else would notice it going missing.
- Containers: the section's authorization handler is registered. Without it
  every container page returns 403 and the suite still passes, since no test
  drives those URLs and the handler's own tests build it by hand.
- Surveys and Consent: section types use their own translation file. Both
  shipped this bug for real - the wrong file gives users raw key names, and
  only on form-failure paths no page test visits.

The same localizer check in the other sections stays deleted; it was template
boilerplate there with no bug behind it.
