+++ tests/Humans.Agent.Tests/AgentArchitectureTests.cs
@@ -1,6 +1,9 @@
 using AwesomeAssertions;
+using Humans.Agent.Contracts;
 using Humans.Agent.Data;
 using Humans.Agent.Domain;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace Humans.Agent.Tests;
 
@@ -56,5 +59,17 @@ public class AgentArchitectureTests
 
         typeof(AgentConversation).GetProperty("UserId").Should().NotBeNull();
         typeof(AgentMessage).GetProperty("HandedOffToFeedbackId").Should().NotBeNull();
+    }
+
+    [HumansFact]
+    public void SectionRegistersTheConversationRetentionForwarder()
+    {
+        // The daily job that deletes expired agent conversations asks for this interface.
+        // Lose the registration and the job stops running, so old conversations are kept
+        // forever. The job itself is not in DI, so no start-up check would notice.
+        var services = new ServiceCollection();
+        new Section().Register(services, new ConfigurationBuilder().Build());
+
+        services.Should().ContainSingle(d => d.ServiceType == typeof(IAgentConversationRetention));
     }
 }
